+++ examples/blink.rs
@@ -1,13 +1,15 @@
 use embedded_hal::digital::v2::OutputPin;
 use ftd2xx_embedded_hal as hal;
-use std::thread::sleep;
-use std::time::Duration;
+use std::{thread::sleep, time::Duration};
 
 const NUM_BLINK: usize = 10;
 const SLEEP_DURATION: Duration = Duration::from_millis(500);
 
 fn main() {
-    let ftdi = hal::Ft232hHal::new().expect("failed to initialize FT232H device");
+    let ftdi = hal::Ft232hHal::new()
+        .expect("Failed to open FT232H device")
+        .init_default()
+        .expect("Failed to initialize MPSSE");
     let mut output_pin = ftdi.ad3();
 
     println!("Starting blinky example");
@@ -16,6 +18,6 @@ fn main() {
         sleep(SLEEP_DURATION);
         output_pin.set_low().expect("failed to set GPIO");
         sleep(SLEEP_DURATION);
-        println!("Blinked {}/{} times", n, NUM_BLINK);
+        println!("Blinked {}/{} times", n + 1, NUM_BLINK);
     }
 }