+++ src/lib.rs
@@ -7,6 +7,11 @@
 //! They have a multi-protocol synchronous serial engine which allows them to
 //! interface with most UART, SPI, and I2C embedded devices.
 //!
+//! **Note:**
+//! This is strictly a development tool.
+//! The crate contains runtime borrow checks and explicit panics to adapt the
+//! FTDI device into the [embedded-hal] traits.
+//!
 //! # Setup
 //!
 //! One-time device setup instructions can be found in the [libftd2xx crate].
@@ -17,10 +22,8 @@
 //!
 //! # Limitations
 //!
-//! * Limited trait support: Blocking SPI and OutputPin traits are implemented.
+//! * Limited trait support: SPI and OutputPin traits are implemented.
 //! * Limited device support: FT232H.
-//! * No methods to control basic parameters such as clock frequency,
-//!   USB latency, timeouts, ect...
 //!
 //! [embedded-hal]: https://crates.io/crates/embedded-hal
 //! [ftdi-embedded-hal]: https://github.com/geomatsi/ftdi-embedded-hal
@@ -34,25 +37,81 @@ pub use embedded_hal;
 pub use libftd2xx;
 
 mod delay;
+mod gpio;
+mod spi;
 
 pub use delay::Delay;
+pub use gpio::OutputPin;
+pub use spi::Spi;
 
-use embedded_hal::spi::Polarity;
-use libftd2xx::{
-    ClockData, ClockDataOut, DeviceTypeError, Ft232h, Ftdi, FtdiCommon, FtdiMpsse, MpsseCmdBuilder,
-    MpsseSettings, TimeoutError,
-};
-use std::cell::RefCell;
-use std::convert::TryFrom;
-use std::sync::Mutex;
+use libftd2xx::{DeviceTypeError, Ft232h, Ftdi, FtdiMpsse, MpsseSettings, TimeoutError};
+use std::{cell::RefCell, convert::TryFrom, sync::Mutex, time::Duration};
 
-/// FT232H device.
-pub struct Ft232hHal {
-    mtx: Mutex<RefCell<Ft232h>>,
-    value: Mutex<u8>,
+/// State tracker for each pin on the FTDI chip.
+#[derive(Debug, Clone, Copy)]
+enum PinUse {
+    MpsseSpi,
+    Output,
 }
 
-impl Ft232hHal {
+struct Ft232hInner {
+    /// FTDI device.
+    ft: Ft232h,
+    /// GPIO direction.
+    direction: u8,
+    /// GPIO value.
+    value: u8,
+    /// Pin allocation.
+    pins: [Option<PinUse>; 8],
+}
+
+impl Ft232hInner {
+    pub(crate) fn with_ftdi(ft: Ft232h) -> Ft232hInner {
+        Ft232hInner {
+            ft,
+            direction: 0xFB,
+            value: 0x00,
+            pins: [None; 8],
+        }
+    }
+
+    /// Allocate a pin for a specific use.
+    pub(crate) fn allocate_pin(&mut self, idx: u8, purpose: PinUse) {
+        assert!(idx < 8, "Pin index {} is out of range 0 - 7", idx);
+
+        if let Some(current) = self.pins[usize::from(idx)] {
+            panic!(
+                "Unable to allocate pin {}, pin is allocated for {:?}",
+                idx, current
+            );
+        } else {
+            self.pins[usize::from(idx)] = Some(purpose)
+        }
+    }
+}
+
+/// Type state for an initialized FTDI HAL.
+///
+/// More information about type states can be found in the [rust-embedded book].
+///
+/// [rust-embedded book]: https://docs.rust-embedded.org/book/static-guarantees/design-contracts.html
+pub struct Initialized;
+
+/// Type state for an uninitialized FTDI HAL.
+///
+/// More information about type states can be found in the [rust-embedded book].
+///
+/// [rust-embedded book]: https://docs.rust-embedded.org/book/static-guarantees/design-contracts.html
+pub struct Uninitialized;
+
+/// FT232H device.
+pub struct Ft232hHal<INITIALIZED> {
+    #[allow(dead_code)]
+    init: INITIALIZED,
+    mtx: Mutex<RefCell<Ft232hInner>>,
+}
+
+impl Ft232hHal<Uninitialized> {
     /// Create a new FT232H structure.
     ///
     /// # Example
@@ -60,10 +119,10 @@ impl Ft232hHal {
     /// ```no_run
     /// use ftd2xx_embedded_hal as hal;
     ///
-    /// let ftdi = hal::Ft232hHal::new()?;
+    /// let ftdi = hal::Ft232hHal::new()?.init_default()?;
     /// # Ok::<(), std::boxed::Box<dyn std::error::Error>>(())
     /// ```
-    pub fn new() -> Result<Ft232hHal, DeviceTypeError> {
+    pub fn new() -> Result<Ft232hHal<Uninitialized>, DeviceTypeError> {
         let ft: Ft232h = Ft232h::try_from(&mut Ftdi::new()?)?;
         Ok(Self::with_ft(ft))
     }
@@ -90,16 +149,96 @@ impl Ft232hHal {
     /// use hal::libftd2xx::Ft232h;
     ///
     /// let ft = Ft232h::with_description("My device description")?;
-    /// let ftdi = hal::Ft232hHal::with_ft(ft);
+    /// let ftdi = hal::Ft232hHal::with_ft(ft).init_default()?;
     /// # Ok::<(), std::boxed::Box<dyn std::error::Error>>(())
     /// ```
-    pub fn with_ft(ft: Ft232h) -> Ft232hHal {
+    pub fn with_ft(ft: Ft232h) -> Ft232hHal<Uninitialized> {
         Ft232hHal {
-            mtx: Mutex::new(RefCell::new(ft)),
-            value: Mutex::new(0x00),
+            init: Uninitialized,
+            mtx: Mutex::new(RefCell::new(Ft232hInner::with_ftdi(ft))),
         }
     }
 
+    /// Initialize the FTDI MPSSE with sane defaults.
+    ///
+    /// Default values:
+    ///
+    /// * Reset the FTDI device.
+    /// * 4k USB transfer size.
+    /// * 1s USB read timeout.
+    /// * 1s USB write timeout.
+    /// * 16ms latency timer.
+    /// * 100kHz clock frequency.
+    ///
+    /// # Example
+    ///
+    /// ```no_run
+    /// use ftd2xx_embedded_hal as hal;
+    /// use hal::{Uninitialized, Initialized, Ft232hHal};
+    ///
+    /// let ftdi: Ft232hHal<Uninitialized> = hal::Ft232hHal::new()?;
+    /// let ftdi: Ft232hHal<Initialized> = ftdi.init_default()?;
+    /// # Ok::<(), std::boxed::Box<dyn std::error::Error>>(())
+    /// ```
+    pub fn init_default(self) -> Result<Ft232hHal<Initialized>, TimeoutError> {
+        const DEFAULT: MpsseSettings = MpsseSettings {
+            reset: true,
+            in_transfer_size: 4096,
+            read_timeout: Duration::from_secs(1),
+            write_timeout: Duration::from_secs(1),
+            latency_timer: Duration::from_millis(16),
+            mask: 0x00,
+            clock_frequency: Some(100_000),
+        };
+
+        self.init(&DEFAULT)
+    }
+
+    /// Initialize the FTDI MPSSE with custom values.
+    ///
+    /// **Note:** The `mask` field of [`MpsseSettings`] is ignored for this function.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the `clock_frequency` field of [`MpsseSettings`] is `None`.
+    ///
+    /// # Example
+    ///
+    /// ```no_run
+    /// use ftd2xx_embedded_hal as hal;
+    /// use hal::{Uninitialized, Initialized, Ft232hHal};
+    /// use hal::libftd2xx::MpsseSettings;
+    ///
+    /// let ftdi: Ft232hHal<Uninitialized> = hal::Ft232hHal::new()?;
+    /// let ftdi: Ft232hHal<Initialized> = ftdi.init(&MpsseSettings {
+    ///     clock_frequency: Some(500_000),
+    ///     ..MpsseSettings::default()
+    /// })?;
+    ///
+    /// # Ok::<(), std::boxed::Box<dyn std::error::Error>>(())
+    /// ```
+    ///
+    /// [`MpsseSettings`]: libftd2xx::MpsseSettings
+    pub fn init(
+        self,
+        mpsse_settings: &MpsseSettings,
+    ) -> Result<Ft232hHal<Initialized>, TimeoutError> {
+        {
+            let lock = self.mtx.lock().expect("Failed to aquire FTDI mutex");
+            let mut inner = lock.borrow_mut();
+            let mut settings = mpsse_settings.clone();
+            settings.mask = inner.direction;
+            inner.ft.initialize_mpsse(&mpsse_settings)?;
+        }
+
+        Ok(Ft232hHal {
+            init: Initialized,
+            mtx: self.mtx,
+        })
+    }
+}
+
+impl Ft232hHal<Initialized> {
     /// Aquire the SPI peripheral for the FT232H.
     ///
     /// Pin assignments:
@@ -107,12 +246,16 @@ impl Ft232hHal {
     /// * AD1 => MOSI
     /// * AD2 => MISO
     ///
+    /// # Panics
+    ///
+    /// Panics if pin 0, 1, or 2 are already in use.
+    ///
     /// # Example
     ///
     /// ```no_run
     /// use ftd2xx_embedded_hal as hal;
     ///
-    /// let ftdi = hal::Ft232hHal::new()?;
+    /// let ftdi = hal::Ft232hHal::new()?.init_default()?;
     /// let spi = ftdi.spi()?;
     /// # Ok::<(), std::boxed::Box<dyn std::error::Error>>(())
     /// ```
@@ -121,203 +264,74 @@ impl Ft232hHal {
     }
 
     /// Aquire the digital output pin 0 for the FT232H.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the pin is already in-use.
     pub fn ad0(&self) -> OutputPin {
-        OutputPin::new(&self.mtx, &self.value, 0)
+        OutputPin::new(&self.mtx, 0)
     }
 
     /// Aquire the digital output pin 1 for the FT232H.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the pin is already in-use.
     pub fn ad1(&self) -> OutputPin {
-        OutputPin::new(&self.mtx, &self.value, 1)
+        OutputPin::new(&self.mtx, 1)
     }
 
     /// Aquire the digital output pin 2 for the FT232H.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the pin is already in-use.
     pub fn ad2(&self) -> OutputPin {
-        OutputPin::new(&self.mtx, &self.value, 2)
+        OutputPin::new(&self.mtx, 2)
     }
 
     /// Aquire the digital output pin 3 for the FT232H.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the pin is already in-use.
     pub fn ad3(&self) -> OutputPin {
-        OutputPin::new(&self.mtx, &self.value, 3)
+        OutputPin::new(&self.mtx, 3)
     }
 
     /// Aquire the digital output pin 4 for the FT232H.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the pin is already in-use.
     pub fn ad4(&self) -> OutputPin {
-        OutputPin::new(&self.mtx, &self.value, 4)
+        OutputPin::new(&self.mtx, 4)
     }
 
     /// Aquire the digital output pin 5 for the FT232H.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the pin is already in-use.
     pub fn ad5(&self) -> OutputPin {
-        OutputPin::new(&self.mtx, &self.value, 5)
+        OutputPin::new(&self.mtx, 5)
     }
 
     /// Aquire the digital output pin 6 for the FT232H.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the pin is already in-use.
     pub fn ad6(&self) -> OutputPin {
-        OutputPin::new(&self.mtx, &self.value, 6)
+        OutputPin::new(&self.mtx, 6)
     }
 
     /// Aquire the digital output pin 7 for the FT232H.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the pin is already in-use.
     pub fn ad7(&self) -> OutputPin {
-        OutputPin::new(&self.mtx, &self.value, 7)
-    }
-}
-
-/// Output pin interface for FTD2xx devices.
-pub struct OutputPin<'a> {
-    mtx: &'a Mutex<RefCell<Ft232h>>,
-    value: &'a Mutex<u8>,
-    mask: u8,
-}
-
-/// SPI interface for FTD2xx devices.
-pub struct Spi<'a> {
-    mtx: &'a Mutex<RefCell<Ft232h>>,
-    clk: ClockData,
-    clk_out: ClockDataOut,
-}
-
-impl<'a> Spi<'a> {
-    pub(crate) fn new(mtx: &Mutex<RefCell<Ft232h>>) -> Result<Spi, TimeoutError> {
-        let lock = mtx.lock().unwrap();
-        let mut ft = lock.borrow_mut();
-        let mut settings = MpsseSettings::default();
-        settings.mask = 0xFB;
-        ft.initialize_mpsse(&settings)?;
-        ft.set_clock(100_000)?;
-        let cmd: MpsseCmdBuilder = MpsseCmdBuilder::new()
-            .set_gpio_lower(0x00, 0xFB)
-            .send_immediate();
-        ft.write_all(cmd.as_slice())?;
-
-        Ok(Spi {
-            mtx,
-            clk: ClockData::MsbPosIn,
-            clk_out: ClockDataOut::MsbNeg,
-        })
-    }
-
-    /// Set the SPI clock polarity.
-    ///
-    /// FTD2XX devices only supports [SPI mode] 0 and 2, clock phase is fixed.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use embedded_hal::spi::Polarity;
-    /// use ftd2xx_embedded_hal as hal;
-    ///
-    /// let ftdi = hal::Ft232hHal::new()?;
-    /// let mut spi = ftdi.spi()?;
-    /// spi.set_clock_polarity(Polarity::IdleLow);
-    /// # Ok::<(), std::boxed::Box<dyn std::error::Error>>(())
-    /// ```
-    ///
-    /// [SPI mode]: https://en.wikipedia.org/wiki/Serial_Peripheral_Interface#Mode_numbers
-    pub fn set_clock_polarity(&mut self, cpol: embedded_hal::spi::Polarity) {
-        let (clk, clk_out) = match cpol {
-            Polarity::IdleLow => (ClockData::MsbPosIn, ClockDataOut::MsbNeg),
-            Polarity::IdleHigh => (ClockData::MsbNegIn, ClockDataOut::MsbPos),
-        };
-
-        // destructuring assignments are unstable
-        self.clk = clk;
-        self.clk_out = clk_out
-    }
-}
-
-impl<'a> embedded_hal::blocking::spi::Write<u8> for Spi<'a> {
-    type Error = TimeoutError;
-    fn write(&mut self, words: &[u8]) -> Result<(), Self::Error> {
-        let cmd: MpsseCmdBuilder = MpsseCmdBuilder::new()
-            .clock_data_out(self.clk_out, words)
-            .send_immediate();
-
-        let lock = self.mtx.lock().unwrap();
-        let mut ft = lock.borrow_mut();
-        ft.write_all(cmd.as_slice())
-    }
-}
-
-impl<'a> embedded_hal::blocking::spi::Transfer<u8> for Spi<'a> {
-    type Error = TimeoutError;
-    fn transfer<'w>(&mut self, words: &'w mut [u8]) -> Result<&'w [u8], Self::Error> {
-        let cmd: MpsseCmdBuilder = MpsseCmdBuilder::new()
-            .clock_data(self.clk, words)
-            .send_immediate();
-
-        let lock = self.mtx.lock().unwrap();
-        let mut ft = lock.borrow_mut();
-        ft.write_all(cmd.as_slice())?;
-        ft.read_all(words)?;
-
-        Ok(words)
-    }
-}
-
-impl<'a> embedded_hal::spi::FullDuplex<u8> for Spi<'a> {
-    type Error = TimeoutError;
-
-    fn read(&mut self) -> nb::Result<u8, Self::Error> {
-        let mut buf: [u8; 1] = [0];
-        let cmd: MpsseCmdBuilder = MpsseCmdBuilder::new()
-            .clock_data(self.clk, &buf)
-            .send_immediate();
-
-        let lock = self.mtx.lock().unwrap();
-        let mut ft = lock.borrow_mut();
-        ft.write_all(cmd.as_slice())?;
-        ft.read_all(&mut buf)?;
-
-        Ok(buf[0])
-    }
-
-    fn send(&mut self, byte: u8) -> nb::Result<(), Self::Error> {
-        let cmd: MpsseCmdBuilder = MpsseCmdBuilder::new()
-            .clock_data_out(self.clk_out, &[byte])
-            .send_immediate();
-
-        let lock = self.mtx.lock().unwrap();
-        let mut ft = lock.borrow_mut();
-        ft.write_all(cmd.as_slice())?;
-        Ok(())
-    }
-}
-
-impl<'a> OutputPin<'a> {
-    pub(crate) fn new(
-        mtx: &'a Mutex<RefCell<Ft232h>>,
-        value: &'a Mutex<u8>,
-        bit: u8,
-    ) -> OutputPin<'a> {
-        let mask: u8 = 1 << bit;
-        OutputPin { mtx, value, mask }
-    }
-
-    pub(crate) fn set(&self, state: bool) -> Result<(), TimeoutError> {
-        let lock = self.mtx.lock().unwrap();
-        let mut value = self.value.lock().unwrap();
-
-        if state {
-            *value |= self.mask;
-        } else {
-            *value &= !self.mask;
-        };
-
-        let mut ft = lock.borrow_mut();
-        let cmd: MpsseCmdBuilder = MpsseCmdBuilder::new()
-            .set_gpio_lower(*value, 0xFB)
-            .send_immediate();
-        ft.write_all(cmd.as_slice())
-    }
-}
-
-impl<'a> embedded_hal::digital::v2::OutputPin for OutputPin<'a> {
-    type Error = TimeoutError;
-
-    fn set_low(&mut self) -> Result<(), Self::Error> {
-        self.set(false)
-    }
-
-    fn set_high(&mut self) -> Result<(), Self::Error> {
-        self.set(true)
+        OutputPin::new(&self.mtx, 7)
     }
 }