+++ CHANGELOG.md
@@ -4,12 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added checks for pin allocation, trying to take output pins 0-2 while using
+  the SPI interface will now result in panic.
+- Added the ability to 
+
 ## [0.4.0] - 2021-03-05
 ### Added
 - Added a `Delay` structure that implements the embedded-hal delay traits.
 
 ### Changed
-- Updated `libftd2xx` dependency from 0.24.0 to 0.25.0.  This updates the vendor library from 1.4.8 to 1.4.22 for Linux targets.  This should fix timeout conditions that previously occurred when rapidly toggling GPIO pins.
+- Updated `libftd2xx` dependency from 0.24.0 to 0.25.0.
+  This updates the vendor library from 1.4.8 to 1.4.22 for Linux targets.
+  This should fix timeout conditions that previously occurred when rapidly toggling GPIO pins.
 
 ## [0.3.0] - 2021-02-14
 ### Changed